+++ src/sorting.c
@@ -160,6 +160,98 @@ merge_sort_merge(void* arr, void* aux, size_t size, int (*compare)(void*, void*)
   }
 }
 
+/**
+ * @brief Sort array of arbitrary values using quicksort.
+ *
+ * If array is small enough, function defers to insertion sort instead.
+ *
+ * @param arr The array to be sorted.
+ * @param nelems The number of elements in the array.
+ * @param size The size of each element in the array.
+ * @param compare The function to be used to compare elements.
+ * @return Void.
+ */
+void
+quick_sort(void* arr, size_t nelems, size_t size, int (*compare)(void*, void*))
+{
+  if (nelems <= LENGTH_THRESHOLD) {
+    insert_sort(arr, nelems, size, compare);
+  } else {
+    quick_sort_sort(arr, size, compare, 1, nelems-1);
+  }
+}
+
+/**
+ * @brief Helper (quick_sort): Recursively perform quicksort.
+ *
+ * Quicksort is not efficient for small arays. As such, insertion sort is used
+ * when subarray is small.
+ *
+ * @param arr The array to be sorted.
+ * @param size The size of each element in the array.
+ * @param compare The function to be used to compare elements.
+ * @param lo The lower index bound of the current subarray.
+ * @param hi The upper index bound of the current subarray.
+ * @return Void.
+ */
+void
+quick_sort_sort(void* arr, size_t size, int (*compare)(void*, void*), size_t lo, size_t hi)
+{
+  if (hi <= lo) {
+    return;
+  } else if (hi - lo <= LENGTH_THRESHOLD) {
+    insert_sort_partial(arr, size, compare, lo, hi+1);
+  } else {
+    size_t pivot = quick_sort_partition(arr, size, compare, lo, hi);
+    quick_sort_sort(arr, size, compare, lo, pivot);
+    quick_sort_sort(arr, size, compare, pivot+1, hi);
+  }
+}
+
+/**
+ * @brief Helper (quick_sort): Partition subarray around pivot element.
+ *
+ * The pivot index is set to the median of the lower, middle, and upper bounds
+ * of the subarray. Choosing the median element results in a faster sort than
+ * always using one of lo, mid, and hi.
+ *
+ * Aside from choice of pivot, this is the Hoare partition scheme.
+ *
+ * @param arr The array containing the subarray.
+ * @param size The size of each element in the array.
+ * @param compare The function to be used to compare elements.
+ * @param lo The lower index bound of the subarray.
+ * @param hi The upper index bound of the subarray.
+ * @return The final index of the pivot element.
+ */
+size_t
+quick_sort_partition(void* arr, size_t size, int (*compare)(void*, void*), size_t lo, size_t hi)
+{
+  char* arr_p = (char*)arr;
+  size_t mid = lo + (hi - lo) / 2;
+  size_t pivot = median_three(arr, size, lo, mid, hi, compare);
+
+  size_t left = lo - 1, right = hi + 1;
+  while (1) {
+    do {
+      left++;
+    } while (compare(arr_p+(left*size), arr_p+(pivot*size)) < 0);
+
+    do {
+      right--;
+    } while (compare(arr_p+(right*size), arr_p+(pivot*size)) > 0);
+
+    if (left >= right) {
+      return right;
+    }
+
+    swap(arr_p+(left*size), arr_p+(right*size), size);
+    if (left == pivot || right == pivot) {
+      pivot = (left == pivot) ? right : left;
+    }
+  }
+}
+
 void
 swap(void* a, void* b, size_t size)
 {
@@ -170,4 +262,41 @@ swap(void* a, void* b, size_t size)
   memcpy(b, tmp, size);
 
   free(tmp);
+}
+
+/**
+ * @brief Find median of three elements in given array and return its index.
+ *
+ * Choosing median of lo, mid, and hi during quicksort gives a better pivot
+ * point than using either a fixed position or the midpoint of hi and lo.
+ *
+ * @param arr The array containing the elements.
+ * @param size The size of each element.
+ * @param a The first element.
+ * @param b The second element.
+ * @param c The third element.
+ * @param compare The function to be used to compare elements.
+ * @return The index of the median element.
+ */
+size_t
+median_three(void* arr, size_t size, size_t a, size_t b, size_t c, int (*compare)(void*, void*))
+{
+  char* arr_p = (char*)arr;
+  if (compare(arr_p+(a*size), arr_p+(b*size)) > 0) {
+    if (compare(arr_p+(b*size), arr_p+(c*size)) > 0) {
+      return b;
+    } else if (compare(arr_p+(a*size), arr_p+(c*size)) > 0) {
+      return c;
+    } else {
+      return a;
+    }
+  } else {
+    if (compare(arr_p+(a*size), arr_p+(c*size)) > 0) {
+      return a;
+    } else if (compare(arr_p+(b*size), arr_p+(c*size)) > 0) {
+      return c;
+    } else {
+      return b;
+    }
+  }
 }