+++ CHANGELOG.md
@@ -14,3 +14,4 @@ The format is based on [Keep a Changelog]() and this project adheres to
 - Insertion sort for both full and partial sorting.
 - Selection sort for full sorting.
 - Merge sort for full sorting.
+- Quicksort for full sorting.